+++ src/appforms/tests/tests/backbone/040-view01form_list.js
@@ -1,16 +1,23 @@
 var assert = chai.assert;
 
 var formListView;
-describe("Backbone - Form List View [test placeholder]", function() {
+var App = {
+    collections: {
+        forms: new Backbone.Collection()
+    }
+};
+describe("Backbone - Form List View", function() {
     it("rendering formView", function(done) {
-        $fh.forms.getForms({fromRemote:false},function(err,forms){
+        $fh.forms.getForms({fromRemote:true},function(err,forms){
             assert (!err);
+
+            var parentEl = new Backbone.View();
             formListView=new FormListView({
                 "model":forms,
-                "parentEl":$("#backbone #formList")
+                "parentEl": parentEl.$el
             });
             formListView.render();
-            assert($("#backbone ul li").length>0, "No Forms rendered");
+            assert(parentEl.$el.find('ul li').length>0, "No Forms rendered");
             done();
         });